requirements.txt
# ============================================================
# requirements.txt — Berlin Snoww Budd Telegram Bot
# Author  : MAZWEWOH JOHN BRINDI
# Python  : >= 3.10
# Updated : 2026
# ============================================================

# Core Telegram bot framework
python-telegram-bot[job-queue]==21.6

# Environment variable loader
python-dotenv==1.0.1

# Async HTTP client (used for image fetching / webhooks / API calls)
httpx==0.27.2

# Telethon – MTProto client (for bot_creator.py — BotFather automation)
# NOTE: Do NOT use Telethon inside the running bot (triggers deletion)
telethon==1.36.0

# Scheduling / task automation
APScheduler==3.10.4

# Logging helpers
loguru==0.7.2

# Utilities
aiofiles==23.2.1
